extract nary-tree/codec pkg
